fix: added readiness probe, grouping handlers
