--- http/healthcheck.go
@@ -1,9 +1,0 @@
-package http
-
-import (
-	"net/http"
-)
-
-func HandleHealthCheck(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(http.StatusOK)
-}
